Refactor report table into multiple components
